Refactor : 각 컴포넌트 style 파일 생성 및 주석 제거
