+++ src/components/Nav/Navbar.scss
@@ -1,0 +1,56 @@
+.header {
+  position: fixed;
+  width: 100%;
+  height: 80px;
+  display: flex;
+  align-items: center;
+  padding: 1.5rem 0;
+  border-bottom: 1px solid rgba(0, 0, 0, 0.4);
+  background-color: #fff;
+  z-index: 10;
+}
+
+.headerContent {
+  max-width: 80%;
+  width: 100%;
+  display: flex;
+  justify-content: space-between;
+  margin: 0 auto;
+}
+
+.logoBox {
+  display: flex;
+  align-items: center;
+  gap: 10px;
+
+  .instagramLogo {
+    height: 25px;
+
+    * {
+      cursor: pointer;
+    }
+  }
+
+  .divider {
+    width: 1px;
+    height: 30px;
+    background-color: #000;
+  }
+
+  .title {
+    font-family: "Lobster", cursive;
+    font-size: 30px;
+    color: black;
+    padding-bottom: 5px;
+  }
+}
+
+.userMenu {
+  display: flex;
+  align-items: center;
+  gap: 40px;
+
+  * {
+    cursor: pointer;
+  }
+}
