+++ src/index.js
@@ -1,6 +1,8 @@
 import React from "react";
 import ReactDOM from "react-dom/client";
 import Router from "./Router";
+import "./style/reset.scss";
+import "./style/common.scss";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(<Router />);
