+++ src/pages/leeinjae/Login/Login.js
@@ -1,17 +1,13 @@
 import React, { useState } from "react";
 import { useNavigate } from "react-router-dom";
-// import { Link } from "react-router-dom";
 import "./Login.scss";
 
 const Login = () => {
-  // const [id, setId] = useState("");
-  // const [password, setPassword] = useState("");
   // useState는 클로저 개념이라 다른 곳에 값을 저장해두고 그것을 가져오는(hooking)하는 느낌
   const [userInfo, setUserInfo] = useState({
     userId: "",
     userPassword: "",
   });
-  // const [isValid, setIsValid] = useState(false);
 
   const navigate = useNavigate();
 
@@ -21,25 +17,6 @@ const Login = () => {
     // 그럴 때 [] 표기법을 사용해 (객체일 때)
     setUserInfo({ ...userInfo, [name]: value });
   };
-
-  // const handleId = (e) => {
-  //   const { value } = e.target;
-  //   setUserInfo({ ...userInfo, id: value });
-  // };
-
-  // const handlePassword = (e) => {
-  //   const { value } = e.target;
-  //   setUserInfo({ ...userInfo, password: value });
-  // };
-
-  // const validation = () => {
-
-  //   console.log(userInfo.userId.includes("@") && userInfo.userPassword.length >= 5)
-
-  //   return userInfo.userId.includes("@") && userInfo.userPassword.length >= 5
-  //     ? setIsValid(true)
-  //     : setIsValid(false);
-  // };
 
   const isValid =
     userInfo.userId.includes("@") && userInfo.userPassword.length >= 5;
