+++ src/pages/leeinjae/Login/Login.scss
@@ -1,6 +1,3 @@
-@import "../../../style/reset.scss";
-@import "../../../style/common.scss";
-
 .container {
   width: 100vw;
   height: 100vh;
