+++ src/pages/leeinjae/Main/components/Feed/Feed.scss
@@ -1,0 +1,130 @@
+.article {
+  border: 1px solid rgba(0, 0, 0, 0.4);
+
+  &:last-child {
+    margin-bottom: 50px;
+  }
+  
+  .writer {
+    max-width: 95%;
+    width: 100%;
+    height: 50px;
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    margin: 0 auto;
+  
+    .userBox {
+      display: flex;
+      align-items: center;
+      gap: 10px;
+  
+      .userProfile {
+        width: 25px;
+        border-radius: 50%;
+      }
+  
+      p {
+        font-weight: bold;
+      }
+    }
+  
+    .menuButton {
+      width: 25px;
+      cursor: pointer;
+    }
+  }
+}
+
+
+.feedImageBox {
+  width: 100%;
+  height: 500px;
+
+  img {
+    width: 100%;
+    height: 100%;
+    object-fit: cover;
+  }
+}
+
+.feedContent {
+  width: 100%;
+  height: auto;
+  padding-bottom: 20px;
+}
+
+.feedTextBox {
+  max-width: 95%;
+  width: 100%;
+  margin: 0 auto;
+  font-size: 14px;
+}
+
+.feedContentImageBox {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  margin: 10px 0;
+
+  .img-left {
+    display: flex;
+    align-items: center;
+    gap: 15px;
+
+    .heart {
+      display: flex;
+      align-items: center;
+    }
+
+    * {
+      width: 25px;
+      cursor: pointer;
+    }
+  }
+
+  .img-right {
+    display: flex;
+    align-items: center;
+
+    * {
+      width: 25px;
+      cursor: pointer;
+    }
+  }
+}
+
+.descriptionBox {
+  display: flex;
+  align-items: center;
+  gap: 5px;
+
+  .descriptionProfile {
+    width: 20px;
+    border-radius: 50%;
+  }
+
+  .bold {
+    font-weight: bold;
+  }
+}
+
+.descriptionContent {
+  margin-top: 20px;
+
+  .initialContent {
+    display: flex;
+    align-items: center;
+    margin-bottom: 10px;
+
+    .bold {
+      font-weight: bold;
+    }
+  }
+
+  .hideContent {
+    display: flex;
+    flex-direction: column;
+    gap: 10px;
+  }
+}
