+++ src/pages/leeinjae/Main/components/Feed/FeedItem.js
@@ -1,12 +1,13 @@
 import React from "react";
+import "./FeedItem.scss";
 
 const FeedItem = ({ lists }) => {
   return (
-    <ul className="hideContent">
+    <>
       {lists.map((elements, index) => (
         <li key={index}>{elements}</li>
       ))}
-    </ul>
+    </>
   );
 };
