+++ src/pages/leeinjae/Main/components/RecommendUser.js
@@ -1,4 +1,5 @@
 import React from "react";
+import "./RecommendUser.scss";
 
 const recommendUserInfo = [
   {
@@ -26,21 +27,21 @@ const recommendUserInfo = [
 
 const RecommendUser = () => {
   return (
-      <div className="userProfile">
-        <div className="recommendUserDescription">
-          {recommendUserInfo.map((el) => (
-            <div className="recommendUserContent" key={el.id}>
-              <img src={el.src} alt={el.alt} />
-              <div className="recommendUserContentText">
-                <p>{el.name}</p>
-                <p>{el.know}</p>
-              </div>
-
-              <button>팔로우</button>
+    <div className="userProfile">
+      <div className="recommendUserDescription">
+        {recommendUserInfo.map((el) => (
+          <div className="recommendUserContent" key={el.id}>
+            <img src={el.src} alt={el.alt} />
+            <div className="recommendUserContentText">
+              <p>{el.name}</p>
+              <p>{el.know}</p>
             </div>
-          ))}
-        </div>
+
+            <button>팔로우</button>
+          </div>
+        ))}
       </div>
+    </div>
   );
 };
