+++ src/pages/leeinjae/Main/components/Search.scss
@@ -1,0 +1,32 @@
+.searchForm {
+  position: relative;
+  display: flex;
+  align-items: center;
+
+  .searchInput {
+    flex: 1;
+    width: 200px;
+    height: 90%;
+    border: 1px solid #000;
+    border-radius: 5px;
+    padding: 0 10px;
+    font-size: 10px;
+
+    &::placeholder {
+      color: #000;
+      text-align: center;
+      font-size: 11px;
+      margin-left: -10px;
+    }
+
+    &:focus::placeholder {
+      opacity: 0;
+    }
+  }
+
+  .searchIcon {
+    position: absolute;
+    left: 80px;
+    z-index: 0;
+  }
+}
