+++ src/pages/leeinjae/Profile/Profile.scss
@@ -1,6 +1,3 @@
-@import "../../../style/reset.scss";
-@import "../../../style/common.scss";
-
 .profileContainer {
   position: relative;
   width: 100vw;
